Move Delivery Retry logic outside EmbraceApiService.
