+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/comms/api/EmbraceApiServiceTest.kt
@@ -3,8 +3,6 @@ package io.embrace.android.embracesdk.comms.api
 import io.embrace.android.embracesdk.ResourceReader
 import io.embrace.android.embracesdk.capture.connectivity.NetworkConnectivityService
 import io.embrace.android.embracesdk.comms.delivery.DeliveryCacheManager
-import io.embrace.android.embracesdk.comms.delivery.DeliveryFailedApiCall
-import io.embrace.android.embracesdk.comms.delivery.DeliveryFailedApiCalls
 import io.embrace.android.embracesdk.comms.delivery.NetworkStatus
 import io.embrace.android.embracesdk.concurrency.BlockingScheduledExecutorService
 import io.embrace.android.embracesdk.config.remote.RemoteConfig
@@ -13,7 +11,6 @@ import io.mockk.clearMocks
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.unmockkAll
-import io.mockk.verify
 import org.junit.After
 import org.junit.AfterClass
 import org.junit.Assert.assertEquals
@@ -24,21 +21,16 @@ import org.junit.Before
 import org.junit.BeforeClass
 import org.junit.Test
 import java.util.concurrent.ScheduledExecutorService
-import java.util.concurrent.TimeUnit
 
 internal class EmbraceApiServiceTest {
 
     companion object {
-        private val connectedNetworkStatuses =
-            NetworkStatus.values().filter { it != NetworkStatus.NOT_REACHABLE }
-
         private lateinit var mockApiUrlBuilder: ApiUrlBuilder
         private lateinit var mockApiClient: ApiClient
         private lateinit var mockCacheManager: DeliveryCacheManager
         private lateinit var blockingScheduledExecutorService: BlockingScheduledExecutorService
         private lateinit var testScheduledExecutor: ScheduledExecutorService
         private lateinit var networkConnectivityService: NetworkConnectivityService
-        private lateinit var failedApiCalls: DeliveryFailedApiCalls
         private lateinit var cachedConfig: CachedConfig
         private lateinit var apiService: EmbraceApiService
 
@@ -50,6 +42,8 @@ internal class EmbraceApiServiceTest {
                 every { getConfigUrl() } returns "https://config.url"
             }
             networkConnectivityService = mockk(relaxUnitFun = true)
+            blockingScheduledExecutorService = BlockingScheduledExecutorService()
+            testScheduledExecutor = blockingScheduledExecutorService
         }
 
         /**
@@ -71,12 +65,7 @@ internal class EmbraceApiServiceTest {
             config = null,
             eTag = null
         )
-        failedApiCalls = DeliveryFailedApiCalls()
-        clearApiPipeline()
         mockCacheManager = mockk(relaxUnitFun = true)
-        every { mockCacheManager.loadPayload("cached_payload_1") } returns "{payload 1}".toByteArray()
-        every { mockCacheManager.loadFailedApiCalls() } returns failedApiCalls
-        every { mockCacheManager.savePayload(any()) } returns "fake_cache"
     }
 
     @After
@@ -94,8 +83,7 @@ internal class EmbraceApiServiceTest {
             headers = emptyMap()
         )
         initApiService(
-            status = NetworkStatus.NOT_REACHABLE,
-            runRetryJobAfterScheduling = true
+            status = NetworkStatus.NOT_REACHABLE
         )
         val remoteConfig = apiService.getConfig()
 
@@ -110,8 +98,7 @@ internal class EmbraceApiServiceTest {
     fun `test getConfig rethrows an exception thrown by apiClient`() {
         every { mockApiClient.executeGet(any()) } throws IllegalStateException("Test exception message")
         initApiService(
-            status = NetworkStatus.NOT_REACHABLE,
-            runRetryJobAfterScheduling = true
+            status = NetworkStatus.NOT_REACHABLE
         )
         // exception will be thrown and caught by this test's annotation
         apiService.getConfig()
@@ -126,213 +113,12 @@ internal class EmbraceApiServiceTest {
             body = "",
             headers = emptyMap()
         )
-        initApiService(
-            status = NetworkStatus.NOT_REACHABLE,
-            runRetryJobAfterScheduling = true
-        )
+        initApiService()
         val remoteConfig = apiService.getConfig()
         assertSame(cfg, remoteConfig)
     }
 
-    @Test
-    fun `scheduled retry job active at init time`() {
-        connectedNetworkStatuses.forEach { status ->
-            initApiService(status = status, runRetryJobAfterScheduling = true)
-            retryTaskActive(status)
-            clearApiPipeline()
-        }
-    }
-
-    @Test
-    fun `retryTask is not active and doesn't run if there are no failed API requests`() {
-        connectedNetworkStatuses.forEach { status ->
-            initApiService(
-                status = status,
-                loadFailedRequest = false,
-                runRetryJobAfterScheduling = true
-            )
-            retryTaskNotActive(status)
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-            checkNoApiRequestSent()
-            retryTaskNotActive(status)
-            clearApiPipeline()
-        }
-    }
-
-    @Test
-    fun `retryTask is active and runs after init if network is connected`() {
-        connectedNetworkStatuses.forEach { status ->
-            initApiService(status = status, runRetryJobAfterScheduling = true)
-            retryTaskActive(status)
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-            checkRequestSendAttempt()
-            retryTaskNotActive(status)
-            clearApiPipeline()
-        }
-    }
-
-    @Test
-    fun `retryTask will be scheduled again if retry fails`() {
-        connectedNetworkStatuses.forEach { status ->
-            every { mockApiClient.post(any(), any()) } throws Exception()
-            initApiService(status = status, runRetryJobAfterScheduling = true)
-            retryTaskActive(status)
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-            checkRequestSendAttempt()
-            retryTaskNotActive(status)
-            // Previous failed attempt will queue another retry. Let it run so a new retry task is active
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-            retryTaskActive(status)
-
-            // First failure will result in another retry in 120 seconds
-            // Go most of the way to check it didn't run
-            blockingScheduledExecutorService.moveForwardAndRunBlocked(
-                TimeUnit.SECONDS.toMillis(119L)
-            )
-            retryTaskActive(status)
-            checkRequestSendAttempt()
-
-            // Go the full 120 seconds and check that the retry runs and fails
-            blockingScheduledExecutorService.moveForwardAndRunBlocked(
-                TimeUnit.SECONDS.toMillis(1L)
-            )
-            checkRequestSendAttempt(count = 2)
-
-            // Previous failed attempt will queue another retry. Let it run
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-
-            // Let the next retry succeed
-            every { mockApiClient.post(any(), any()) } returns ""
-
-            // Second failure will result in another retry in double the last time, 240 seconds
-            // Go most of the way to check it didn't run, then go all the way to check that it did.
-            blockingScheduledExecutorService.moveForwardAndRunBlocked(
-                TimeUnit.SECONDS.toMillis(239L)
-            )
-            retryTaskActive(status)
-            checkRequestSendAttempt(count = 2)
-            blockingScheduledExecutorService.moveForwardAndRunBlocked(
-                TimeUnit.SECONDS.toMillis(1L)
-            )
-            retryTaskNotActive(status)
-            checkRequestSendAttempt(count = 3)
-            clearApiPipeline()
-        }
-    }
-
-    @Test
-    fun `retryTask is not active and doesn't run after init if network not reachable`() {
-        initApiService(
-            status = NetworkStatus.NOT_REACHABLE,
-            runRetryJobAfterScheduling = true
-        )
-        blockingScheduledExecutorService.runCurrentlyBlocked()
-        retryTaskNotActive(NetworkStatus.NOT_REACHABLE)
-        checkNoApiRequestSent()
-    }
-
-    @Test
-    fun `retryTask isn't active and won't run if there are no failed requests after getting a connection before retry job is scheduled`() {
-        connectedNetworkStatuses.forEach { status ->
-            initApiService(
-                status = NetworkStatus.NOT_REACHABLE,
-                loadFailedRequest = false,
-                runRetryJobAfterScheduling = true
-            )
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-            apiService.onNetworkConnectivityStatusChanged(status)
-            retryTaskNotActive(status)
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-            checkNoApiRequestSent()
-            retryTaskNotActive(status)
-            clearApiPipeline()
-        }
-    }
-
-    @Test
-    fun `retryTask is active and runs after connection changes from not reachable to connected after retry job runs`() {
-        connectedNetworkStatuses.forEach { status ->
-            initApiService(
-                status = NetworkStatus.NOT_REACHABLE,
-                runRetryJobAfterScheduling = true
-            )
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-            apiService.onNetworkConnectivityStatusChanged(status)
-            retryTaskActive(status)
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-            checkRequestSendAttempt()
-            retryTaskNotActive(status)
-            clearApiPipeline()
-        }
-    }
-
-    @Test
-    fun `retryTask isn't active and doesn't run if there are no failed request after getting a connection before retry job is scheduled`() {
-        connectedNetworkStatuses.forEach { status ->
-            initApiService(
-                status = NetworkStatus.NOT_REACHABLE,
-                loadFailedRequest = false
-            )
-            apiService.onNetworkConnectivityStatusChanged(status)
-            retryTaskNotActive(status)
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-            checkNoApiRequestSent()
-            retryTaskNotActive(status)
-            clearApiPipeline()
-        }
-    }
-
-    @Test
-    fun `retryTask is active and runs after connection changes from not reachable to connected before retry job is scheduled`() {
-        connectedNetworkStatuses.forEach { status ->
-            initApiService(status = NetworkStatus.NOT_REACHABLE)
-            apiService.onNetworkConnectivityStatusChanged(status)
-            retryTaskActive(status)
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-            checkRequestSendAttempt()
-            retryTaskNotActive(status)
-            clearApiPipeline()
-        }
-    }
-
-    @Test
-    fun `retryTask is not active and doesn't run after connection changes from connected to not reachable before retry job is scheduled`() {
-        connectedNetworkStatuses.forEach { status ->
-            initApiService(status = status)
-            apiService.onNetworkConnectivityStatusChanged(NetworkStatus.NOT_REACHABLE)
-            retryTaskNotActive(status)
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-            checkNoApiRequestSent()
-        }
-    }
-
-    @Test
-    fun `queue size should be bounded`() {
-        initApiService(status = NetworkStatus.WIFI, loadFailedRequest = false)
-        every { mockApiClient.post(any(), any()) } throws Exception()
-
-        assertEquals(0, apiService.pendingRetriesCount())
-
-        repeat(201) {
-            apiService.sendSession("{ dummy_session }".toByteArray(), null)
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-        }
-        assertEquals(200, apiService.pendingRetriesCount())
-    }
-
-    private fun clearApiPipeline() {
-        clearMocks(mockApiClient, answers = false)
-        failedApiCalls.clear()
-        blockingScheduledExecutorService = BlockingScheduledExecutorService()
-        testScheduledExecutor =
-            blockingScheduledExecutorService
-    }
-
-    private fun initApiService(
-        status: NetworkStatus,
-        loadFailedRequest: Boolean = true,
-        runRetryJobAfterScheduling: Boolean = false
-    ) {
+    private fun initApiService(status: NetworkStatus = NetworkStatus.NOT_REACHABLE) {
         every { networkConnectivityService.getCurrentNetworkStatus() } returns status
 
         apiService = EmbraceApiService(
@@ -346,34 +132,7 @@ internal class EmbraceApiServiceTest {
             cacheManager = mockCacheManager,
             lazyDeviceId = lazy { "07D85B44E4E245F4A30E559BFC0D07FF" },
             appId = "o0o0o",
+            deliveryRetryManager = mockk(relaxed = true)
         )
-
-        failedApiCalls.clear()
-        if (loadFailedRequest) {
-            failedApiCalls.add(DeliveryFailedApiCall(mockk(), "cached_payload_1"))
-        }
-
-        if (runRetryJobAfterScheduling) {
-            blockingScheduledExecutorService.runCurrentlyBlocked()
-        }
-    }
-
-    private fun retryTaskActive(status: NetworkStatus) {
-        assertTrue("Failed for network status = $status", apiService.isRetryTaskActive())
-    }
-
-    private fun retryTaskNotActive(status: NetworkStatus) {
-        assertFalse(
-            "Failed for network status = $status",
-            apiService.isRetryTaskActive()
-        )
-    }
-
-    private fun checkRequestSendAttempt(count: Int = 1) {
-        verify(exactly = count) { mockApiClient.post(any(), "{payload 1}".toByteArray()) }
-    }
-
-    private fun checkNoApiRequestSent() {
-        verify(exactly = 0) { mockApiClient.post(any(), any()) }
     }
 }
