+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/injection/FakeEssentialServiceModule.kt
@@ -13,6 +13,7 @@ import io.embrace.android.embracesdk.comms.api.ApiService
 import io.embrace.android.embracesdk.comms.api.ApiUrlBuilder
 import io.embrace.android.embracesdk.comms.delivery.CacheService
 import io.embrace.android.embracesdk.comms.delivery.DeliveryCacheManager
+import io.embrace.android.embracesdk.comms.delivery.DeliveryRetryManager
 import io.embrace.android.embracesdk.config.ConfigService
 import io.embrace.android.embracesdk.fakes.FakeActivityService
 import io.embrace.android.embracesdk.fakes.FakeAndroidMetadataService
@@ -22,6 +23,7 @@ import io.embrace.android.embracesdk.fakes.FakeCacheService
 import io.embrace.android.embracesdk.fakes.FakeConfigService
 import io.embrace.android.embracesdk.fakes.FakeCpuInfoDelegate
 import io.embrace.android.embracesdk.fakes.FakeDeliveryCacheManager
+import io.embrace.android.embracesdk.fakes.FakeDeliveryRetryManager
 import io.embrace.android.embracesdk.fakes.FakeDeviceArchitecture
 import io.embrace.android.embracesdk.fakes.FakeGatingService
 import io.embrace.android.embracesdk.fakes.FakeMemoryCleanerService
@@ -51,6 +53,7 @@ internal class FakeEssentialServiceModule(
     override val networkConnectivityService: NetworkConnectivityService = NoOpNetworkConnectivityService(),
     override val cacheService: CacheService = FakeCacheService(),
     override val deliveryCacheManager: DeliveryCacheManager = FakeDeliveryCacheManager(),
+    override val deliveryRetryManager: DeliveryRetryManager = FakeDeliveryRetryManager(),
     override val urlBuilder: ApiUrlBuilder = FakeApiUrlBuilder()
 ) : EssentialServiceModule {
